requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0  # Add this line
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
python-decouple==3.8
email-validator==2.1.0
slowapi==0.1.9
beanie==1.23.6
bcrypt==4.0.1
aiosmtplib